+++ http.go
@@ -1,8 +1,9 @@
 package sola
 
 import (
-	"net/http"
 	"sync"
+
+	"github.com/fvbock/endless"
 )
 
 var wg sync.WaitGroup
@@ -12,7 +13,7 @@ func Listen(addr string, g *Sola) {
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		http.ListenAndServe(addr, g)
+		endless.ListenAndServe(addr, g)
 	}()
 }
 
@@ -21,7 +22,7 @@ func ListenTLS(addr string, certFile string, keyFile string, g *Sola) {
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		http.ListenAndServeTLS(addr, certFile, keyFile, g)
+		endless.ListenAndServeTLS(addr, certFile, keyFile, g)
 	}()
 }